Separate schema loader and code generator.
