+++ GeneratorSchema/GeneratorSchema.csproj
@@ -2,6 +2,7 @@
 
 	<PropertyGroup>
 		<TargetFramework>netstandard2.0</TargetFramework>
+		<RootNamespace>KhronosGroup.Gltf.Generator.Schema</RootNamespace>
 	</PropertyGroup>
 
 	<ItemGroup>
