+++ GeneratorUnitTests/GetCodegenTypeTests.cs
@@ -1,9 +1,12 @@
-﻿using System;
+using System;
 using System.CodeDom;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using GeneratorLib;
+
+using KhronosGroup.Gltf.Generator.Schema;
+
 using Newtonsoft.Json.Linq;
 using NUnit.Framework;
